added the hydra client
