+++ client-side/src/app/app.module.ts
@@ -24,7 +24,7 @@ import { TraineeTechSkillsComponent } from './portals/Caliber/reports/trainee-te
 import { HttpClientModule } from '@angular/common/http';
 import { RoleGuard } from './role-guard';
 import { CookieService } from 'ngx-cookie-service';
-import { HydraClientModule } from './portals/hydra-client/hydra-client.module';
+import { HydraClientModule } from './hydra-client/hydra-client.module';
 
 // loading routes from child modules this way will lazy load them
 const routes: Routes = [
