+++ client-side/src/app/hydra-client/category/hydra-category.service.spec.ts
@@ -1,0 +1,15 @@
+import { TestBed, inject } from '@angular/core/testing';
+
+import { HydraCategoryService } from './hydra-category.service';
+
+describe('HydraCategoryService', () => {
+  beforeEach(() => {
+    TestBed.configureTestingModule({
+      providers: [HydraCategoryService]
+    });
+  });
+
+  it('should be created', inject([HydraCategoryService], (service: HydraCategoryService) => {
+    expect(service).toBeTruthy();
+  }));
+});
