+++ client-side/src/app/hydra-client/category/hydra-category.service.ts
@@ -1,0 +1,9 @@
+import { Injectable } from '@angular/core';
+
+@Injectable()
+export class HydraCategoryService {
+  test = 'hydra test';
+
+  constructor() { }
+
+}
