+++ client-side/src/app/hydra-client/hydra-client.module.ts
@@ -1,0 +1,13 @@
+import { NgModule } from '@angular/core';
+import { HttpModule } from '@angular/http';
+import { HydraCategoryService } from './category/hydra-category.service';
+
+@NgModule({
+  imports: [
+    HttpModule,
+  ],
+  declarations: [
+  ],
+  providers: [HydraCategoryService]
+})
+export class HydraClientModule { }
