+++ client-side/src/app/portals/Caliber/home/home.component.ts
@@ -1,4 +1,5 @@
 import { Component, OnInit, OnDestroy } from '@angular/core';
+import { HydraCategoryService } from '../../../hydra-client/category/hydra-category.service';
 
 @Component({
   selector: 'app-home',
@@ -6,8 +7,12 @@ import { Component, OnInit, OnDestroy } from '@angular/core';
   styleUrls: ['./home.component.css']
 })
 export class HomeComponent implements OnInit, OnDestroy {
-  constructor() { }
 
-  ngOnInit() {}
+  constructor(private hydraCategoryService: HydraCategoryService) { }
+
+  ngOnInit() {
+    alert(this.hydraCategoryService.test);
+  }
+
   ngOnDestroy() {}
 }
