--- client-side/src/app/portals/Hydra-Client/hydra-client.module.ts
@@ -1,12 +1,0 @@
-import { NgModule } from '@angular/core';
-import { CommonModule } from '@angular/common';
-import { HttpClientModule } from '@angular/common/http';
-
-@NgModule({
-  imports: [
-    CommonModule,
-    HttpClientModule
-  ],
-  declarations: []
-})
-export class HydraClientModule { }
