--- client-side/src/app/portals/Hydra-Client/services/url/url.service.spec.ts
@@ -1,15 +1,0 @@
-import { TestBed, inject } from '@angular/core/testing';
-
-import { UrlService } from './url.service';
-
-describe('UrlService', () => {
-  beforeEach(() => {
-    TestBed.configureTestingModule({
-      providers: [UrlService]
-    });
-  });
-
-  it('should be created', inject([UrlService], (service: UrlService) => {
-    expect(service).toBeTruthy();
-  }));
-});
